Attempt to establish operational baseline
Initial configuration set up for linux on x86_64. based on i386 linux configuration.
Large numbers of potential bounds check failures detected at compile-time.
51 of 55 systems tests crash.
The original archived removed asl (and this is probably why), leaving behind only dangling links - this is a copy restored from an external source.
This copy might also be removed; since the original source is trying to force-fit C into being Pascal and is not native C at all.
Later versions of asl exist ... which might be integrated into this version., if it is kept.